add authorize for policy to role permission module

+++ Modules/RolePermission/Http/Controllers/RolePermissionController.php
@@ -2,8 +2,13 @@
 
 namespace Modules\RolePermission\Http\Controllers;
 
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Contracts\Foundation\Application;
+use Illuminate\Contracts\View\Factory;
+use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
 use Modules\RolePermission\Http\Requests\RolePermissionRequest;
+use Modules\RolePermission\Models\Permission;
 use Modules\RolePermission\Repositories\RolePermissionRepo;
 use Modules\RolePermission\Services\RolePermissionService;
 use Modules\Share\Http\Controllers\Controller;
@@ -12,6 +17,8 @@ use Modules\Share\Services\ShareService;
 
 class RolePermissionController extends Controller
 {
+    private string $class = Permission::class;
+
     public RolePermissionRepo $repo;
     public RolePermissionService $service;
 
@@ -25,19 +32,27 @@ class RolePermissionController extends Controller
      * Get latest roles.
      *
      * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
+     * @throws AuthorizationException
      */
     public function index()
     {
+        $this->authorize('manage', $this->class);
         $roles = $this->repo->index()->paginate(10);
+
         return view('RolePermission::index', compact('roles'));
     }
 
     /**
      * Create page for role.
+     *
+     * @return Application|Factory|View
+     * @throws AuthorizationException
      */
     public function create()
     {
+        $this->authorize('manage', $this->class);
         $permissions = $this->repo->getAllPermissions();
+
         return view('RolePermission::create', compact('permissions'));
     }
 
@@ -46,9 +61,11 @@ class RolePermissionController extends Controller
      *
      * @param  RolePermissionRequest $request
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function store(RolePermissionRequest $request)
     {
+        $this->authorize('manage', $this->class);
         $this->service->store($request);
 
         return $this->successMessageWithRedirect('Create role');
@@ -59,9 +76,11 @@ class RolePermissionController extends Controller
      *
      * @param  $id
      * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View
+     * @throws AuthorizationException
      */
     public function edit($id)
     {
+        $this->authorize('manage', $this->class);
         $role = $this->repo->findById($id);
         $permissions = $this->repo->getAllPermissions();
 
@@ -71,12 +90,14 @@ class RolePermissionController extends Controller
     /**
      * Update role by id.
      *
-     * @param  RolePermissionRequest $request
+     * @param RolePermissionRequest $request
      * @param  $id
      * @return RedirectResponse
+     * @throws AuthorizationException
      */
     public function update(RolePermissionRequest $request, $id)
     {
+        $this->authorize('manage', $this->class);
         $this->service->update($request, $id);
 
         return $this->successMessageWithRedirect('Update role');
@@ -87,10 +108,13 @@ class RolePermissionController extends Controller
      *
      * @param  $id
      * @return \Illuminate\Http\JsonResponse
+     * @throws AuthorizationException
      */
     public function destroy($id)
     {
+        $this->authorize('manage', $this->class);
         $this->repo->delete($id);
+
         return AjaxResponses::SuccessResponse();
     }
